Add patch to make servo pulses more accurate
Originally all interrupts were at the same priority. This meant
that when an interrupt performed some calculation while the
servo pulse interrupt Timer1 triggered, the servo pulse
interrupt was delayed and therefore the servo pulse longer
than desired. This was found to be about 16us -- which is
quite significant.
We now configure the Timer1 interrupt to have higher priority,
which means it will be running even if another interrupt is active.

More stack is required, it has not been verified yet that sufficient
stack space is left.