Make the crystal oscillator work for 8ch version
The PLL is not used in the 8ch version.
The 4ch version has changed to use the post divider of the PLL
instead of the AHB divider by 4.
This way the clock that the UART sees is the same in all cases.

This also affects the IOCON module, to be checked if we need
to change anything.